pyproject.toml
[project]
name = "video-balancer"
version = "0.1.0"
description = ""
authors = [
    {name = "asaWasa"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "sqlalchemy[asyncio] (>=2.0.0,<3.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "uvicorn[standard] (>=0.30.0,<0.31.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "colorlog (>=6.8.0,<7.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
